Adding support for the :jar-exlusions option.
* note that leiningen specifies[1] the use of a regular expression
for this property, while currently it is being used as an ant
selector pattern.  If maintaining this strict compatibility is important
than I recommend adding support for the filename selector[2] in uncle.  Rather
than expecting a string pattern for inclusion and exclusion patterns it could
check for a vector or strings and/or a regular expressions.  When a regular
expression is found it should use a nested filename selector element with the regex
property rather than a normal file pattern.

In the meantime, this lets us exclude files from the jar (e.g. the user.clj,
which should maybe be excluded by default anyways.)

[1] https://github.com/technomancy/leiningen/blob/stable/sample.project.clj
[2] http://ant.apache.org/manual/Types/selectors.html#filenameselect